techpack: dsi: Force setting brightness after AOD OFF
The screen's backlight will not turn on if we do the following things by
step in a very dark room:
  1. Turn on AOD display mode
  2. Turn off the screen (AOD will on)
  3. Cover the top of the screen (The screen will turn off)
  4. Remove the cover (AOD will on)
  5. Turn on the screen (!! The screen won't turn on !!)

Seems that the brightness of the screen will be messed up (Don't know
exactly what's going on though). So force setting brightness when
turning the AOD OFF, so that the backlight will work properly.

Change-Id: If56b359b51b69be07da78aa4413271c5cdf05add